# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Freelancer Weekly Planner API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "404":
          description: No profile yet
    post:
      operationId: createProfile
      tags: [profile]
      summary: Create or replace user profile (onboarding)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
    patch:
      operationId: updateProfile
      tags: [profile]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"

  # Clients
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks (optionally filter by week)
      parameters:
        - name: weekStart
          in: query
          required: false
          schema:
            type: string
            description: ISO date for the start of the week (YYYY-MM-DD)
        - name: clientId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Task list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Income entries
  /income:
    get:
      operationId: listIncome
      tags: [income]
      summary: List income entries
      parameters:
        - name: month
          in: query
          required: false
          schema:
            type: string
            description: YYYY-MM
      responses:
        "200":
          description: Income list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IncomeEntry"
    post:
      operationId: createIncome
      tags: [income]
      summary: Record an income entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncomeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncomeEntry"
  /income/{id}:
    patch:
      operationId: updateIncome
      tags: [income]
      summary: Update an income entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncomeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncomeEntry"
    delete:
      operationId: deleteIncome
      tags: [income]
      summary: Delete an income entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Income Goals
  /income-goals:
    get:
      operationId: listIncomeGoals
      tags: [income]
      summary: List income goals
      responses:
        "200":
          description: Income goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IncomeGoal"
    post:
      operationId: createIncomeGoal
      tags: [income]
      summary: Create or update an income goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncomeGoalInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncomeGoal"
  /income-goals/{id}:
    patch:
      operationId: updateIncomeGoal
      tags: [income]
      summary: Update an income goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncomeGoalUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncomeGoal"
    delete:
      operationId: deleteIncomeGoal
      tags: [income]
      summary: Delete an income goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Expenses
...
components:
  schemas:
...
    Profile:
      type: object
      required: [id, name, profession, monthlGoal, weeklyHours, currency, niche, workDays]
      properties:
        id:
          type: integer
        name:
          type: string
        profession:
          type: string
        monthlyGoal:
          type: number
        weeklyHours:
          type: number
        currency:
          type: string
        niche:
          type: string
        workDays:
          type: array
          items:
            type: string
        timezone:
          type: ["string", "null"]
        weekStartDay:
          type: string
        darkMode:
          type: boolean
    ProfileInput:
      type: object
      required: [name, profession, monthlyGoal, weeklyHours, currency, niche, workDays]
      properties:
        name:
          type: string
        profession:
          type: string
        monthlyGoal:
          type: number
        weeklyHours:
          type: number
        currency:
          type: string
        niche:
          type: string
        workDays:
          type: array
          items:
            type: string
        timezone:
          type: string
        weekStartDay:
          type: string
        darkMode:
          type: boolean
    ProfileUpdate:
      type: object
      properties:
        name:
          type: string
        profession:
          type: string
        monthlyGoal:
          type: number
        weeklyHours:
          type: number
        currency:
          type: string
        niche:
          type: string
        workDays:
          type: array
          items:
            type: string
        timezone:
          type: string
        weekStartDay:
          type: string
        darkMode:
          type: boolean
...
    Task:
      type: object
      required: [id, title, status, priority, scheduledDate]
      properties:
        id:
          type: integer
        title:
          type: string
        clientId:
          type: ["integer", "null"]
        clientName:
          type: ["string", "null"]
        deadline:
          type: ["string", "null"]
        scheduledDate:
          type: string
          description: ISO date YYYY-MM-DD
        estimatedHours:
          type: ["number", "null"]
        actualHours:
          type: ["number", "null"]
        priority:
          type: string
          description: "low | medium | high | urgent"
        status:
          type: string
          description: "todo | in_progress | waiting | delivered | revision | completed"
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    TaskInput:
      type: object
      required: [title, status, priority, scheduledDate]
      properties:
        title:
          type: string
        clientId:
          type: integer
        deadline:
          type: string
        scheduledDate:
          type: string
        estimatedHours:
          type: number
        actualHours:
          type: number
        priority:
          type: string
        status:
          type: string
        notes:
          type: string
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        clientId:
          type: integer
        deadline:
          type: string
        scheduledDate:
          type: string
        estimatedHours:
          type: number
        actualHours:
          type: number
        priority:
          type: string
        status:
          type: string
        notes:
          type: string
    IncomeEntry:
      type: object
      required: [id, amount, currency, date, description]
      properties:
        id:
          type: integer
        amount:
          type: number
        currency:
          type: string
        date:
          type: string
        description:
          type: string
        clientId:
          type: ["integer", "null"]
        clientName:
          type: ["string", "null"]
        invoiceNumber:
          type: ["string", "null"]
        paymentStatus:
          type: string
          description: "paid | pending | overdue"
        createdAt:
          type: string
    IncomeInput:
      type: object
      required: [amount, currency, date, description]
      properties:
        amount:
          type: number
        currency:
          type: string
        date:
          type: string
        description:
          type: string
        clientId:
          type: integer
        invoiceNumber:
          type: string
        paymentStatus:
          type: string
    IncomeUpdate:
      type: object
      properties:
        amount:
          type: number
        currency:
          type: string
        date:
          type: string
        description:
          type: string
        clientId:
          type: integer
        invoiceNumber:
          type: string
        paymentStatus:
          type: string
    IncomeGoal:
      type: object
      required: [id, period, targetAmount, currency]
      properties:
        id:
          type: integer
        period:
          type: string
          description: "weekly | monthly | quarterly"
        targetAmount:
          type: number
        currency:
          type: string
        currentAmount:
          type: number
        periodLabel:
          type: string
        createdAt:
          type: string
    IncomeGoalInput:
      type: object
      required: [period, targetAmount, currency]
      properties:
        period:
          type: string
        targetAmount:
          type: number
        currency:
          type: string
        periodLabel:
          type: string
    IncomeGoalUpdate:
      type: object
      properties:
        period:
          type: string
        targetAmount:
          type: number
        currency:
          type: string
        periodLabel:
          type: string
...