# lib/api-zod/src/generated/types/incomeEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Freelancer Weekly Planner API
 * OpenAPI spec version: 0.1.0
 */

export interface IncomeEntry {
  id: number;
  amount: number;
  currency: string;
  date: string;
  description: string;
  /** @nullable */
  clientId?: number | null;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  invoiceNumber?: string | null;
  /** paid | pending | overdue */
  paymentStatus?: string;
  createdAt?: string;
}