# lib/api-zod/src/generated/types/incomeUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Freelancer Weekly Planner API
 * OpenAPI spec version: 0.1.0
 */

export interface IncomeUpdate {
  amount?: number;
  currency?: string;
  date?: string;
  description?: string;
  clientId?: number;
  invoiceNumber?: string;
  paymentStatus?: string;
}