# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Freelancer Weekly Planner API
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  id: number;
  name: string;
  profession: string;
  monthlyGoal?: number;
  weeklyHours: number;
  currency: string;
  niche: string;
  workDays: string[];
  /** @nullable */
  timezone?: string | null;
  weekStartDay?: string;
  darkMode?: boolean;
}