# lib/api-zod/src/generated/types/profileUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Freelancer Weekly Planner API
 * OpenAPI spec version: 0.1.0
 */

export interface ProfileUpdate {
  name?: string;
  profession?: string;
  monthlyGoal?: number;
  weeklyHours?: number;
  currency?: string;
  niche?: string;
  workDays?: string[];
  timezone?: string;
  weekStartDay?: string;
  darkMode?: boolean;
}