# lib/api-zod/src/generated/types/taskInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Freelancer Weekly Planner API
 * OpenAPI spec version: 0.1.0
 */

export interface TaskInput {
  title: string;
  clientId?: number;
  deadline?: string;
  scheduledDate: string;
  estimatedHours?: number;
  actualHours?: number;
  priority: string;
  status: string;
  notes?: string;
}